v. 2.5.3 check all fields

+++ src/main/java/com/course_project/support/TrainManager.java
@@ -40,6 +40,9 @@ public class TrainManager extends Manager implements UpdatableCountWagons {
     }
 
     public boolean deleteTrain(String nameTrain) {
+        if (trainDao.findByName(nameTrain) == null ) {
+            return false;
+        }
         train.setName(nameTrain);
         return trainDao.deleteByName(train);
     }
